Fixed bug where routing tables were updated to worst value instead of best value...
